requirements.txt
tensorflow==2.15.0
streamlit
opencv-python
pillow
numpy==1.26.4
matplotlib
scikit-learn